code pipeline
I don't want to train all the model over and over again so from the notebook.
it is clear that which model is doing best.
so i will write pipeline code in this notebook.
which train and test that specifict model from notebook.
The best performing model from notebook is - random forest clssification.

use `test.py` for testing the model